requirements.txt
PyQt5
PyOpenGL
PyOpenGL_accelerate
networkx
pandas
